// src/app/modules/models/materiaTexto.ts
export class MateriaTexto {
    mattexid: any;
    matid: any;
    matnombre: any;
    texid: any;
    texnombre: any;
    textipo: any;
    texdocumento: any;
    mattexdescripcion: any;
    mattexusureg: any;
    mattexfecreg: any;
    mattexusumod: any;
    mattexfecmod: any;
    mattexestado: any;
    tipo: any;
}